+++ src/main/java/com/job/application/tracker/service/UserService.java
@@ -5,6 +5,7 @@ import com.job.application.tracker.dto.UserCreateDto;
 import com.job.application.tracker.dto.UserGetDto;
 import com.job.application.tracker.dto.UserUpdateDto;
 import com.job.application.tracker.entity.User;
+import com.job.application.tracker.exceptions.ResourceNotFoundException;
 import com.job.application.tracker.mapper.UserMapper;
 import com.job.application.tracker.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -40,18 +41,22 @@ public class UserService implements IUserService {
 
     @Override
     public UserGetDto get(Integer id) {
-        User user = userRepository.findById(id).orElseThrow();
+        User user = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
         return UserMapper.toDto(user);
     }
 
     @Override
     public void delete(Integer id) {
+        if(!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User not found with id:" + id);
+        }
+
         userRepository.deleteById(id);
     }
 
     @Override
     public UserGetDto update(Integer id ,UserUpdateDto userDto) {
-        User user = userRepository.findById(id).orElseThrow();
+        User user = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User not found with id: " + id));
          UserMapper.UpdateEntity(user , userDto);
         userRepository.save(user);
         return UserMapper.toDto(user);
